+++ CHANGELOG.md
@@ -23,7 +23,9 @@
 * [Breaking] Rename `PopulateStruct` method in value to `Populate`.
   The method can now populate not only structs, but anything: slices,
   maps, builtin types and maps.
-  
+* [Breaking] Pass a tracer the `uhttp/uhttpclient` constructor explicitly, instead
+  of using a global tracer. This will allow to use http client in parallel tests.
+
 ## v1.0.0-beta2 (09 Mar 2017)
 
 * [Breaking] Remove `ulog.Logger` interface and expose `*zap.Logger` directly.